Further reduced dependencies of varx_rx.cpp
