+++ varx/rx/internal/varx_Observer_Impl.cpp
