+++ varx/rx/internal/varx_Observer_Impl.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace detail {
+    struct ObserverImpl
+    {
+        ObserverImpl(any&& wrapped);
+        void onNext(any&& next) const;
+        void onError(Error error) const;
+        void onCompleted() const;
+        
+        const any wrapped;
+    };
+}
