+++ varx/rx/varx_Observer.h
@@ -1,17 +1,5 @@
 #pragma once
 
-namespace detail {
-    struct ObserverImpl
-    {
-        ObserverImpl(any&& wrapped);
-        void onNext(any&& next) const;
-        void onError(Error error) const;
-        void onCompleted() const;
-        
-        const any wrapped;
-    };
-}
-
 /**
  Retrieves items. You can call onNext to notify the Observer with a new item.
