+++ varx/varx.h
@@ -36,26 +36,36 @@ END_JUCE_MODULE_DECLARATION
 #include "util/internal/concurrentqueue.h"
 
 namespace varx {
-    typedef std::exception_ptr Error;
-    
-    struct Empty {};
-    inline bool operator==(const Empty& lhs, const Empty& rhs){ return true; }
-    inline bool operator!=(const Empty& lhs, const Empty& rhs){ return !(lhs == rhs); }
+typedef std::exception_ptr Error;
+
+struct Empty
+{};
+inline bool operator==(const Empty& lhs, const Empty& rhs)
+{
+    return true;
+}
+inline bool operator!=(const Empty& lhs, const Empty& rhs)
+{
+    return !(lhs == rhs);
+}
 }
 
 namespace varx {
 #include "util/internal/varx_any.h"
 #include "rx/varx_Disposable.h"
 #include "rx/varx_DisposeBag.h"
+#include "rx/internal/varx_Observer_Impl.h"
 #include "rx/varx_Observer.h"
+#include "rx/internal/varx_Observable_Impl.h"
 #include "rx/varx_Observable.h"
 #include "rx/varx_Scheduler.h"
+#include "rx/internal/varx_Subjects_Impl.h"
 #include "rx/varx_Subjects.h"
-    
+
 #include "util/internal/varx_ReleasePool.h"
 #include "util/varx_LockFreeSource.h"
 #include "util/varx_LockFreeTarget.h"
-    
+
 #include "integration/varx_ExtensionBase.h"
 #include "integration/varx_GUIExtensions.h"
 #include "integration/varx_ModelExtensions.h"
@@ -71,7 +81,11 @@ VARX_DEFINE_VARIANT_CONVERTER(varx::detail::ObservableImpl)
 VARX_DEFINE_VARIANT_CONVERTER(varx::Empty)
 
 template<typename... Args>
-struct juce::VariantConverter<std::tuple<Args...>> : varx::detail::VariantConverter<std::tuple<Args...>> {};
+struct juce::VariantConverter<std::tuple<Args...>> : varx::detail::VariantConverter<std::tuple<Args...>>
+{
+};
 
 template<typename T>
-struct juce::VariantConverter<varx::Observable<T>> : varx::detail::VariantConverter<varx::Observable<T>> {};
+struct juce::VariantConverter<varx::Observable<T>> : varx::detail::VariantConverter<varx::Observable<T>>
+{
+};
