+++ varx/varx_rx.cpp
@@ -14,17 +14,17 @@ typedef std::exception_ptr Error;
 #include "rx/varx_DisposeBag.h"
 #include "rx/internal/varx_Disposable_Impl.h"
 #include "rx/internal/varx_DisposeBag_Impl.h"
-#include "rx/varx_Observable.h"
+#include "rx/internal/varx_Observable_Impl.h"
 #include "rx/varx_Scheduler.h"
-#include "rx/varx_Observer.h"
+#include "rx/internal/varx_Observer_Impl.h"
 #include "rx/internal/varx_Scheduler_Impl.h"
-#include "rx/varx_Subjects.h"
+#include "rx/internal/varx_Subjects_Impl.h"
 #include "rx/varx_Disposable.cpp"
 #include "rx/varx_DisposeBag.cpp"
-#include "rx/varx_Observable.cpp"
-#include "rx/varx_Observer.cpp"
 #include "rx/varx_Scheduler.cpp"
-#include "rx/varx_Subjects.cpp"
 #include "rx/internal/varx_Disposable_Impl.cpp"
 #include "rx/internal/varx_Scheduler_Impl.cpp"
+#include "rx/internal/varx_Observable_Impl.cpp"
+#include "rx/internal/varx_Observer_Impl.cpp"
+#include "rx/internal/varx_Subjects_Impl.cpp"
 }
