--- varx/varx_util.cpp
@@ -1,11 +1,0 @@
-#include "JuceHeader.h"
-
-namespace varx {
-using namespace juce;
-
-#include "util/internal/varx_any.cpp"
-#include "util/internal/varx_ReleasePool.cpp"
-#include "util/varx_LockFreeSource.cpp"
-#include "util/varx_PrintFunctions.cpp"
-#include "util/varx_VariantConverters.cpp"
-}
